+++ driverless_ws/src/perceptions/setup.py
@@ -20,9 +20,9 @@ setup(
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
-            'data_node = perceptions.utils.DataNode:main',
-            'stereo_node = perceptions.predict.StereoNode:main',
-            'lidar_node = perceptions.predict.LidarNode:main'
+            'data_node = perceptions.DataNode:main',
+            'stereo_node = perceptions.StereoNode:main',
+            'lidar_node = perceptions.LidarNode:main'
         ],
     },
 )
